task: APP_NAME => APP_CODE, no runtime config (#5)
As it turns out: publishRuntimeConfig and serverRuntimeConfig options
are deprecated in the latest version of Next. The docs recommend to just
use environment variables. Added bonus is that any environment variable
that is not prefixed with NEXT_PUBLIC_ will only exist on the server and
not be exposed to the client