Task: reorg packages for better dx (#50)
Changes here are mostly cosmetic/organizational but with real benefits
once this codebase starts to get larger.

The gist of this change is that I've reorganized code into domain
packages rather than generic folders like `store` or `api`. I was
noticing that we were starting to get some merge conflicts which would
only get worse as those generic package names start to grow and everyone
is working in them. For example, we now have a `user` package that deals
with all user related functionality. Rather than user functionality
being spread across multiple packages like `store` and `api`, it now all
lives under `packages/user`.

The issue with keeping the generic package names is that it's very easy
to be lazy and just dump a new API function inside of the `api` package
alongside everything else. It requires no thought and leads to things
like `data.js` that we have in our existing app. A more intentional
package structure requires a bit more thought up front on where to place
new functionality but pays off in the long run.

I also consolidated some packages and was able to delete some packages
as a result.