EHR-6723: fix middlewares (#208)
The gist: we were doing a lot of strange things to get next to play
nicely with the nginx ingress and trying to maintain different logic
based on the environment i.e. we were using the ingress paths in all
deployed environments and not using in in development. Upon digging, I
found a Nextjs configuration called `basePath` that essentially is made
for the purpose of when your apps are hosted on the same domain at
different paths (like we do). This greatly simplifies the code and
removes the need for any hacks, although we will now have to use the
path name in development as well (not a big deal)

Having different behavior in local vs prod caused a long-existing bug to
be exposed. The requests being made by the clinical widget was actually
being routed through the patient apps middleware! The reason why it was
never caught was because all apps had identical middleware. Now, with
the new patient app, that is no longer the case (each dedicated app will
have its own custom middleware for its own specific needs). When we
replaced the patient app middleware, it caused the other apps to break
because they were not making API requests using their base paths. Adding
`basePath` config resolves this, now each app will be scoped to its own
middleware as intended